net/trash/gpt_scraper_v2/requirements.txt
# ChatGPT Analyzer - Python Requirements
# Production dependencies for all features

# Core dependencies
selenium>=4.15.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.2
requests>=2.31.0
lxml>=4.9.3

# Environment variable support
python-dotenv>=1.0.0

# CAPTCHA handling
opencv-python>=4.8.1.78
Pillow>=10.1.0
numpy>=1.24.4

# GUI automation (optional - may have issues on some systems)
# pyautogui>=0.9.54

# OCR support (optional but recommended)
pytesseract>=0.3.10

# PDF generation (optional)
reportlab>=4.0.7
PyPDF2>=3.0.1

# Data processing
pandas>=2.1.3
openpyxl>=3.1.2  # For Excel export

# Browser automation helpers
undetected-chromedriver>=3.5.4  # Helps avoid detection
selenium-stealth>=1.0.6

# Utilities
colorama>=0.4.6  # Cross-platform colored output
tqdm>=4.66.1  # Progress bars
click>=8.1.7  # CLI enhancements

# Development dependencies (optional)
# Uncomment these for development work
# ipython>=8.17.2
# black>=23.11.0
# flake8>=6.1.0
# pytest>=7.4.3
# pytest-cov>=4.1.0
# mypy>=1.7.0

# Platform-specific notes:
# Windows users may need:
# - Visual C++ Build Tools for some packages
# - May need to install Tesseract OCR separately
#
# Linux users may need:
# - sudo apt-get install tesseract-ocr (Debian/Ubuntu)
# - sudo yum install tesseract (RedHat/CentOS)
# - For pyautogui: sudo apt-get install python3-tk python3-dev
#
# macOS users may need:
# - brew install tesseract (via Homebrew)
#
# For Chromium on Ubuntu:
# - sudo apt-get install chromium-chromedriver
# - or: sudo snap install chromium
